Crate parser-combinators

This crate contains parser combinators, roughly based on the Haskell library parsec.

A parser in this library can be described as a function which takes some input and if it is succesful, returns a value together with the remaining input. A parser combinator is a function which takes one or more parsers and returns a new parser. For instance the many parser can be used to convert a parser for single digits into one that parses multiple digits.

Examples:

 extern crate "parser-combinators" as parser_combinators;
 use parser_combinators::{spaces, many1, sep_by, digit, satisfy, Parser, ParserExt, ParseError};
 
 fn main() {
     let input = "1234, 45,78";
     let spaces = spaces();
     let integer = spaces.clone()//Parse spaces first and use the with method to only keep the result of the next parser
         .with(many1(digit()).map(|string: String| string.parse::<i32>().unwrap()));//parse a string of digits into an i32
     //Parse integers separated by commas, skipping whitespace
     let mut integer_list = sep_by(integer, spaces.skip(satisfy(|c| c == ',')));
 
     //Call parse with the input to execute the parser
     let result: Result<(Vec<i32>, &str), ParseError> = integer_list.parse(input);
     match result {
         Ok((value, _remaining_input)) => println!("{:?}", value),
         Err(err) => println!("{}", err)
     }
 }

If we need a parser that is mutually recursive we can define a free function which internally can in turn be used as a parser (Note that we need to explicitly cast the function, this should not be necessary once changes in rustc to make orphan checking less restrictive gets implemented)

expr is written fully general here which may not be necessary in a specific implementation The Stream trait is predefined to work with array slices, string slices and iterators meaning that in this case it could be defined as fn expr(input: State<&str>) -> ParseResult<Expr, &str>

 extern crate "parser-combinators" as parser_combinators;
 use parser_combinators::{between, spaces, many1, parser, sep_by, satisfy, Parser, ParserExt,
 ParseResult};
 use parser_combinators::primitives::{State, Stream};

 #[derive(Debug, PartialEq)]
 enum Expr {
     Id(String),
     Array(Vec<Expr>),
 }
 fn expr<I>(input: State<I>) -> ParseResult<Expr, I>
     where I: Stream<Item=char> {
     let word = many1(satisfy(|c| c.is_alphabetic()));
     let comma_list = sep_by(parser(expr), satisfy(|c| c == ','));
     let array = between(satisfy(|c| c == '['), satisfy(|c| c == ']'), comma_list);
     let spaces = spaces();
     spaces.clone().with(
             word.map(Expr::Id)
             .or(array.map(Expr::Array))
         ).parse_state(input)
 }
 
 fn main() {
     let result = parser(expr)
         .parse("[[], hello, [world]]");
     let expr = Expr::Array(vec![
           Expr::Array(Vec::new())
         , Expr::Id("hello".to_string())
         , Expr::Array(vec![Expr::Id("world".to_string())])
     ]);
     assert_eq!(result, Ok((expr, "")));
 }

Modules

char
Module containg parsers specialized on character streams

combinator
Module containing all specific parsers

primitives
Module containing the primitive types which is used to create and compose more advanced parsers

Structs

ParseError
Struct which hold information about an error that occured at a specific position. Can hold multiple instances of Error if more that one error occured at the position.

Traits

Parser
By implementing the Parser trait a type says that it can be used to parse an input stream into the type Output.

ParserExt
Extension trait which provides functions that are more conveniently used through method calls

Functions

alpha_num
Parses either an alphabet letter or digit

any_char
Parses any character

between
Parses open followed by parser followed by close Returns the value of parser

chainl1
Parses p 1 or more times separated by op The value returned is the one produced by the left associative application of op

choice_slice

choice_vec

crlf
Parses carriage return and newline, returning the newline character.

digit
Parses a digit from a stream containing characters

hex_digit
Parses a hexdecimal digit with uppercase and lowercase

letter
Parses an alphabet letter

lower
Parses an uppercase letter

many
Parses p zero or more times returning a collection with the values from p. If the returned collection cannot be inferred type annotations must be supplied, either by annotating the resulting type binding let collection: Vec<_> = ... or by specializing when calling many, many::<Vec<_>, _>(...)

many1
Parses p one or more times returning a collection with the values from p. If the returned collection cannot be inferred type annotations must be supplied, either by annotating the resulting type binding let collection: Vec<_> = ... or by specializing when calling many1 many1::<Vec<_>, _>(...)

newline
Parses a newline character

not_followed_by
Succeeds only if parser fails. Never consumes any input.

oct_digit
Parses an octal digit

optional
Returns Some(value) and None on parse failure (always succeeds)

parser
Wraps a function, turning it into a parser Mainly needed to turn closures into parsers as function types can be casted to function pointers to make them usable as a parser

satisfy
Parses a character and succeeds depending on the result of pred

sep_by
Parses parser zero or more time separated by separator, returning a collection with the values from p. If the returned collection cannot be inferred type annotations must be supplied, either by annotating the resulting type binding let collection: Vec<_> = ... or by specializing when calling sep_by, sep_by::<Vec<_>, _, _>(...)

skip_many
Parses p zero or more times ignoring the result

skip_many1
Parses p one or more times ignoring the result

space
Parses whitespace

spaces
Skips over zero or more spaces

string
Parses the string s

tab
Parses a tab character

try
Try acts as p except it acts as if the parser hadn't consumed any input if p returns an error after consuming input

unexpected
Always fails with message as the error. Never consumes any input.

upper
Parses an uppercase letter

value
Always returns the value v without consuming any input.

Type Definitions

ParseResult
A type alias over the specific Result type used to indicated parser success/failure. O is the type that is output on success I is the specific stream type used in the parser