Module parser-combinators::char

Module containg parsers specialized on character streams

Structs

AlphaNum

CrLf

Digit

HexDigit

Letter

Lower

NewLine

OctDigit

Satisfy

Space

Spaces

String

Tab

Upper

Functions

alpha_num
Parses either an alphabet letter or digit

any_char
Parses any character

crlf
Parses carriage return and newline, returning the newline character.

digit
Parses a digit from a stream containing characters

hex_digit
Parses a hexdecimal digit with uppercase and lowercase

letter
Parses an alphabet letter

lower
Parses an uppercase letter

newline
Parses a newline character

oct_digit
Parses an octal digit

satisfy
Parses a character and succeeds depending on the result of pred

space
Parses whitespace

spaces
Skips over zero or more spaces

string
Parses the string s

tab
Parses a tab character

upper
Parses an uppercase letter