Function parser_combinators::chainl1

pub fn chainl1<'a, P, Op>(parser: P, op: Op) -> Chainl1<P, Op> where P: Parser, Op: Parser<Input=P::Input, Output=Box<FnMut(P::Output, P::Output) -> P::Output + 'a>>

Parses p 1 or more times separated by op The value returned is the one produced by the left associative application of op