Function parser-combinators::many1

pub fn many1<F, P>(p: P) -> Many1<F, P> where F: FromIterator<<P as Parser>::Output>, P: Parser

Parses p one or more times returning a collection with the values from p. If the returned collection cannot be inferred type annotations must be supplied, either by annotating the resulting type binding let collection: Vec<_> = ... or by specializing when calling many1 many1::<Vec<_>, _>(...)

 let result = many1::<Vec<_>, _>(digit())
     .parse("A123");
 assert!(result.is_err());