Module parser-combinators::primitives

Module containing the primitive types which is used to create and compose more advanced parsers

Structs

ParseError
Struct which hold information about an error that occured at a specific position. Can hold multiple instances of Error if more that one error occured at the position.

SourcePosition
Struct which containing the current position

State
The State<I> struct keeps track of the current position in the stream I

Enums

Consumed
Enum used to indicate if a stream has had any elements consumed

Error
Enum used to store information about an error that has occured

Traits

Parser
By implementing the Parser trait a type says that it can be used to parse an input stream into the type Output.

Stream
A stream is a sequence of items that can be extracted one by one

Type Definitions

ParseResult
A type alias over the specific Result type used to indicated parser success/failure. O is the type that is output on success I is the specific stream type used in the parser