Enum parser_combinators::primitives::Consumed

pub enum Consumed<T> {
    Consumed(T),
    Empty(T),
}

Enum used to indicate if a parser consumed any items of the stream it was given as an input

Variants

Consumed

Constructor indicating that the parser has consumed elements

Empty

Constructor indicating that the parser did not consume any elements

Methods

impl<T> Consumed<T>

fn is_empty(&self) -> bool

Returns true if self is empty

fn into_inner(self) -> T

Extracts the contained value

fn as_consumed(self) -> Consumed<T>

Converts self into the Consumed state

fn as_empty(self) -> Consumed<T>

Converts self into theEmpty state

fn map<F, U>(self, f: F) -> Consumed<U> where F: FnOnce(T) -> U

Maps over the contained value without changing the consumed state

fn combine<F, U, I>(self, f: F) -> ParseResult<U, I, I::Item> where F: FnOnce(T) -> ParseResult<U, I, I::Item>, I: Stream

Combines the Consumed flags from self and the result of f

 //Parses a characther of string literal and handles the escaped characthers \\ and \" as \
 //and " respectively
 fn char(input: State<&str>) -> ParseResult<char, &str> {
     let (c, input) = try!(satisfy(|c| c != '"').parse_state(input));
     match c {
         //Since the `char` parser has already consumed some of the input `combine` is used
         //propagate the consumed state to the next part of the parser
         '\\' => input.combine(|input| {
             satisfy(|c| c == '"' || c == '\\')
                 .map(|c| {
                     match c {
                         '"' => '"',
                         '\\' => '\\',
                         c => c
                     }
                 })
                 .parse_state(input)
             }),
         _ => Ok((c, input))
     }
 }
 let result = many(parser(char))
     .parse(r#"abc\"\\"#);
 assert_eq!(result, Ok((r#"abc"\"#.to_string(), "")));
 }