Enum parser_combinators::primitives::Error

pub enum Error<T> {
    Unexpected(T),
    Expected(Info<T>),
    Message(Info<T>),
    Other(Box<StdError + Send>),
}

Enum used to store information about an error that has occured

Variants

Unexpected

Error indicating an unexpected token has been encountered in the stream

Expected

Error indicating that the parser expected something else

Message

Generic message

Other

Variant for containing other types of errors