Struct parser_combinators::primitives::State

pub struct State<I> where I: Stream {
    pub position: I::Item::Position,
    pub input: I,
}

The State<I> struct keeps track of the current position in the stream I

Fields

position

input

Methods

impl<I: Stream> State<I>

fn new(input: I) -> State<I>

fn as_empty(&self) -> State<I>

fn uncons(self) -> ParseResult<I::Item, I, I::Item>

uncons is the most general way of extracting and item from a stream It takes a function f as argument which should update the position according to the item that was extracted Usually you want to use uncons_char instead which works directly on character streams