Struct parser_combinators::ParseError

pub struct ParseError<P: Positioner> {
    pub position: P::Position,
    pub errors: Vec<Error<P>>,
}

Struct which hold information about an error that occured at a specific position. Can hold multiple instances of Error if more that one error occured at the position.

Fields

position

The position where the error occured

errors

A vector containing specific information on what errors occured at position

Methods

impl<P: Positioner> ParseError<P>

fn new(position: P::Position, error: Error<P>) -> ParseError<P>

fn from_errors(position: P::Position, errors: Vec<Error<P>>) -> ParseError<P>

fn add_message<S>(&mut self, message: S) where S: Into<Info<P>>

fn add_error(&mut self, message: Error<P>)

fn set_expected(&mut self, message: Info<P>)

fn merge(self, other: ParseError<P>) -> ParseError<P>