Enum parser-combinators::primitives::Consumed

pub enum Consumed<T> {
    Consumed(T),
    Empty(T),
}

Enum used to indicate if a stream has had any elements consumed

Variants

Consumed

Constructor indicating that the parser has consumed elements

Empty

Constructor indicating that the parser did not consume any elements

Methods

impl<T> Consumed<T>

fn is_empty(&self) -> bool

Returns true if the self is empty

fn into_inner(self) -> T

Extracts the contained value

fn as_consumed(self) -> Consumed<T>

Converts the consumed state into the Consumed state

fn as_empty(self) -> Consumed<T>

Converts the consumed state into the Empty state

fn map<F, U>(self, f: F) -> Consumed<U> where F: FnOnce(T) -> U

Maps over the contained value without changing the consumed state

fn combine<F, U, I>(self, f: F) -> ParseResult<U, I> where F: FnOnce(T) -> ParseResult<U, I>

Combines the Consumed flags from self and the result of f