Struct parser-combinators::primitives::State

pub struct State<I> {
    pub position: SourcePosition,
    pub input: I,
}

The State<I> struct keeps track of the current position in the stream I

Fields

position

input

Methods

impl<I: Stream> State<I>

fn new(input: I) -> State<I>

fn as_empty(&self) -> State<I>

fn uncons<F>(self, f: F) -> ParseResult<<I as Stream>::Item, I> where F: FnOnce(&mut SourcePosition, &<I as Stream>::Item)

uncons is the most general way of extracting and item from a stream It takes a function f as argument which should update the position according to the item that was extracted Usually you want to use uncons_char instead which works directly on character streams

impl<I: Stream<Item=char>> State<I>

fn uncons_char(self) -> ParseResult<<I as Stream>::Item, I>

Specialized uncons function for character streams which updates the position with no further action needed