Struct parser-combinators::ParseError

pub struct ParseError {
    pub position: SourcePosition,
    pub errors: Vec<Error>,
}

Struct which hold information about an error that occured at a specific position. Can hold multiple instances of Error if more that one error occured at the position.

Fields

position

The position where the error occured

errors

A vector containing specific information on what errors occured at position

Methods

impl ParseError

fn new(position: SourcePosition, error: Error) -> ParseError

fn add_message<S>(&mut self, message: S) where S: IntoCow<'static, str>

fn add_error(&mut self, message: Error)

fn set_expected(&mut self, message: Cow<'static, str>)

fn merge(self, other: ParseError) -> ParseError